securebank/requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.2
django-apscheduler==0.7.0
django-otp==1.5.4
django-password-validation==0.1.1
django-ranged-response==0.2.0
django-rest-framework==0.1.0
django-simple-captcha==0.6.0
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
pytz==2024.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
zope.interface==7.1.0
